Add documentation to user interface methods
This commit adds documentation to the user interface methods, including
brief explanations of some implementation choices (mainly for when we
do not include certain cases).

This commit also fixes some small errors. I forgot to include the IDLE
state case for two of the event handlers, so I added it.