Integrate user interface with screens and latest wave detection (#5)
* Add template for event transition handlers

This commit adds the method headers for the event transition handlers, including the basic switch statements for each event case. The methods are currently unimplemented.

* Implement event transition handler methods

This commit implements the event transition handler methods for
ONE_PERSON, MULTIPLE_PEOPLE, ONE_PERSON_WAVING to each other state.

* Add event handler functionality for TWO_PEOPLE_WAVING, GENERATING, and PRESENTING states

* Add documentation to user interface methods

This commit adds documentation to the user interface methods, including brief explanations of some implementation choices (mainly for when we do not include certain cases).

This commit also fixes some small errors. I forgot to include the IDLE state case for two of the event handlers, so I added it.

* Fix event handlers

This commit adds the event transition between ONE_PERSON_WAVING and TWO_PEOPLE_WAVING. This commit also deletes the event handler for TWO_PEOPLE_WAVING and calls the PRESENTING event handler when it switches to that state.

* Integrate screens with user interface

* Introduce animations between screen components in UI

---------

Co-authored-by: jadekeegan <97476936+jadekeegan@users.noreply.github.com>
Co-authored-by: briannata <brianna.hy.ta@gmail.com>